Get chalk printing working again.
