+++ lib/elm-test.js
@@ -22,11 +22,9 @@ process.on("uncaughtException", function(error) {
 });
 
 var compile = require("node-elm-compiler").compile,
-  os = require("os"),
   fs = require("fs-extra"),
   crypto = require("crypto"),
   glob = require("glob"),
-  chalk = require("chalk"),
   builder = require("xmlbuilder"),
   path = require("path"),
   util = require("util"),
@@ -36,6 +34,7 @@ var compile = require("node-elm-compiler").compile,
   firstline = require("firstline"),
   chokidar = require("chokidar"),
   Runner = require("./runner.js"),
+  Supervisor = require("./supervisor.js"),
   Init = require("./init.js"),
   child_process = require("child_process");
 
@@ -202,101 +201,7 @@ function runTests(testFile) {
       }
       prepareCompiledJsFile(initialSeed, report, dest);
 
-      function finish() {
-        // TODO take code from readAndEval to print summary and exit.
-        console.log("FINSIHED!");
-      }
-
-      var cpus = os.cpus() || new Array(1);
-
-      var nextTestToRun = -1;
-      var finishedWorkers = 0;
-      var closedWorkers = 0;
-
-      function runNextTest(worker) {
-        var testToRun = nextTestToRun;
-
-        nextTestToRun++;
-
-        // Immediately run the next test.
-        worker.send("" + testToRun);
-      }
-
-      var workers = cpus.map(function(throwaway, index) {
-        var worker = child_process.fork(dest);
-
-        worker.on("close", function(code, signal) {
-          closedWorkers++;
-
-          // If all the workers have closed, we're done!
-          if (closedWorkers === workers.length) {
-            finish();
-          }
-        });
-
-        worker.on("message", function(data) {
-          console.log("[WORKER " + index + "]", data);
-
-          var response = JSON.parse(data);
-          switch (response.type) {
-            case "FINISHED":
-              // This worker found no tests remaining to run; it's finished!
-              finishedWorkers++;
-
-              // If all the workers have finished, print the summmary.
-              if (finishedWorkers === workers.length) {
-                // TODO send enough summary data that one worker can do this.
-                worker.send("SUMMARY");
-              }
-              break;
-            case "SUMMARY":
-              // TODO print the summary.
-              console.log("Received SUMMARY");
-
-              // Close all the workers.
-              workers.forEach(function(worker) {
-                worker.kill();
-              });
-              break;
-            case "BEGIN":
-              // TODO record all the relevant values, and display the "hey we're
-              // starting up here" thing to the user.
-              console.log("Received BEGIN");
-
-              runNextTest(worker);
-
-              break;
-            default:
-              // TODO record the result!
-              // TODO print progress bar - e.g. "Running test 5 of 20" on a bar!
-              // -- yikes, be careful though...test the scenario where test
-              // authors put Debug.log in their tests - does that mess
-              // everything up re: the line feed? Seems like it would...
-              // ...so maybe a bar is not best. Can we do better? Hm.
-
-              runNextTest(worker);
-          }
-        });
-
-        return worker;
-      });
-
-      // Set the workers running.
-      workers.forEach(function(worker, index) {
-        var testToRun = nextTestToRun;
-
-        nextTestToRun++;
-
-        if (testToRun === -1) {
-          // The BEGIN message requests metadata about the test run, e.g.
-          // how many tests will be run, whether they should auto-fail because
-          // of skip/on/y/todo, etc.
-          worker.send("BEGIN");
-        } else {
-          // Send the index of the test to run.
-          worker.send("" + testToRun);
-        }
-      });
+      Supervisor.run(dest);
     }
   });
 }
@@ -325,34 +230,6 @@ function prepareCompiledJsFile(initialSeed, report, dest) {
   ].join("\n");
 
   fs.writeFileSync(dest, finalContent);
-}
-
-function readAndEval(dest) {
-  fs.readFile(dest, { encoding: "utf8" }, function(readErr, compiledElmCode) {
-    if (readErr) {
-      console.error(
-        "The test run failed because it could not read the compiled Elm code:\n\n",
-        readErr
-      );
-
-      if (!args.watch) {
-        process.exit(1);
-      }
-    }
-
-    try {
-      Runner.evalElmCode(args, report, compiledElmCode);
-    } catch (err) {
-      console.error(
-        "The test run failed because of a runtime exception encountered when evaluating the compiled Elm code:\n\n",
-        err
-      );
-
-      if (!args.watch) {
-        process.exit(1);
-      }
-    }
-  });
 }
 
 function checkNodeVersion() {
