+++ lib/runner.js
@@ -89,132 +89,6 @@ function findNearestElmPackageDir(filePaths /*:Array<string>*/) {
   return process.cwd();
 }
 
-function evalElmCode(
-  args /*:Array<string>*/,
-  report /*:string*/,
-  compiledCode /*:string*/
-) {
-  // Apply Node polyfills as necessary.
-  var window = {
-    Date: Date,
-    addEventListener: function() {},
-    removeEventListener: function() {}
-  };
-  var document = { body: {}, createTextNode: function() {} };
-  if (typeof XMLHttpRequest === "undefined") {
-    XMLHttpRequest = function() {
-      return {
-        addEventListener: function() {},
-        open: function() {},
-        send: function() {}
-      };
-    };
-  }
-  if (typeof FormData === "undefined") {
-    FormData = function() {
-      this._data = [];
-    };
-    FormData.prototype.append = function() {
-      this._data.push(Array.prototype.slice.call(arguments));
-    };
-  }
-
-  var Elm = (function(module) {
-    eval(compiledCode);
-    return module.exports;
-  })({});
-
-  // Make sure necessary things are defined.
-  if (typeof Elm === "undefined") {
-    throw "test runner config error: Elm is not defined. Make sure you provide a file compiled by Elm!";
-  }
-
-  var potentialModuleNames = Object.keys(Elm.Test.Generated);
-
-  if (potentialModuleNames.length !== 1) {
-    console.error(
-      "Multiple potential generated modules to run in the Elm.Test.Generated namespace: ",
-      potentialModuleNames,
-      " - this should never happen!"
-    );
-    process.exit(1);
-  }
-
-  var testModule = Elm.Test.Generated[potentialModuleNames[0]];
-  var initialSeed = null;
-
-  if (args.seed !== undefined) {
-    initialSeed = args.seed;
-  }
-
-  // Fix Windows Unicode problems. Credit to https://github.com/sindresorhus/figures for the Windows compat idea!
-  var windowsSubstitutions = [
-    [/[↓✗►]/g, ">"],
-    [/╵│╷╹┃╻/g, "|"],
-    [/═/g, "="],
-    ,
-    [/▔/g, "-"],
-    [/✔/g, "√"]
-  ];
-
-  function windowsify(str) {
-    return windowsSubstitutions.reduce(function(result, sub) {
-      return result.replace(sub[0], sub[1]);
-    }, str);
-  }
-
-  function chalkify(messages) {
-    return messages
-      .map(function(msg) {
-        var path = msg.styles;
-        var text =
-          process.platform === "win32" ? windowsify(msg.text) : msg.text;
-
-        if (path.length === 0) {
-          return text;
-        } else {
-          var fn = chalk;
-
-          path.forEach(function(nextPath) {
-            fn = fn[nextPath];
-          });
-
-          return fn(text);
-        }
-      })
-      .join("");
-  }
-
-  // Run the Elm app.
-  var app = testModule.worker({ seed: initialSeed, report: report });
-
-  // // Receive messages from ports and translate them into appropriate JS calls.
-  app.ports.send.subscribe(function(msg) {
-    // var msgType = msg[0];
-    // var data = msg[1];
-    //
-    // if (msgType === "FINISHED") {
-    //   if (data.format === "CHALK") {
-    //     console.log(chalkify(data.message));
-    //   } else if (data.format === "JUNIT") {
-    //     console.log(builder.create(data.message).end());
-    //   } else {
-    //     console.log(JSON.stringify(data.message));
-    //   }
-    //
-    //   if (!args.watch) {
-    //     process.exit(data.exitCode);
-    //   }
-    // } else if (msgType === "STARTED" || msgType === "TEST_COMPLETED") {
-    //   if (data.format === "CHALK") {
-    //     console.log(chalkify(data.message));
-    //   } else if (data.format === "JSON") {
-    //     console.log(JSON.stringify(data.message));
-    //   }
-    // }
-  });
-}
-
 var binaryExtension = process.platform === "win32" ? ".exe" : "";
 var readElmiPath =
   path.join(__dirname, "..", "bin", "elm-interface-to-json") + binaryExtension;
@@ -428,7 +302,6 @@ function filterExposing(pathAndModule) {
 
 module.exports = {
   findTests: findTests,
-  evalElmCode: evalElmCode,
   isElmFile: isElmFile,
   copyNativeSrcFiles: copyNativeSrcFiles,
   repositoryToNativePackageName: repositoryToNativePackageName,
